requirements.txt
python-telegram-bot==20.3
pandas
gspread
psycopg2-binary
psycopg2
